10.0.18
It looks like automatic network discovery is causing issues so it was removed, I also need to update the freeform toggle so that you can use 32:9 without black bars.
This update includes a 1% opacity on the close and dimming buttons, and then they will (on mouse hover) go to 100% opacity on mouse hover.
Hopefully this reduces the glare